Fiddling with firebird (to be ditched).
I've decided to ditch any attempts to work with Firebird (2.5). It's
just too much of a pain in the ass for now, whether configuring users
or creating databases. But the cherry on top was how much work it is
just to create an autoincrementing primary key with a sequence.

Rather than having a bit of syntactic sugar that would let you do
something like this:

    CREATE SEQUENCE myseq;

    CREATE TABLE widgets (
        widget_id INTEGER NOT NULL DEFAULT NEXT_SEQ(myseq) PRIMARY KEY
    );

You have to go through some big rigamarole with a trigger _just_ to
generate the next item in the sequence.

Now, much as I'm a fan of using natural and compound keys as my primary
keys where possible, there are some circumstances in which that's not
particularly useful, and I have two tables in a simple three-table
schema where using an autoincrementing primary key makes the most sense.
That means I have to use PostgreSQL, SQLite, or _shudder_ MySQL.

Currently, I'm using SQLite for the table, and I was intending on using
this small project to see how easy/painful Firebird would be. From this
perspective, it's just not worth it yet. Maybe when Firebird 3 is
available for FreeBSD I'll rethink my decision.