Configure cgit.
This is enough to just about make cgit useful.

I want to use git-http-backend for git push, but that will require me to
add some thing of auth mechanism first, which I currently don't have.

I'm painted into a corner as far as that goes because I have to rely on
the nginx http_auth_request module[1], which will require some coding to
support. I could build a custom package with additional modules
activated, or use nginx-full, but in the latter case, I'd might as well
use Apache, and in the former, it's almost as much of a headache to
support.

Another alternative would be a git SSH user, and something like
gitolite[2] or gitosis[3] to do user and repo management. Those are
'fun' in their own way. Of the two, the latter is preferable, but I need
to get the port updated to 0.3 first.

[1]: http://nginx.org/en/docs/http/ngx_http_auth_request_module.html
[2]: http://gitolite.com/gitolite/
[3]: https://github.com/tv42/gitosis